backend/requirements.txt
# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation and settings
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# LangChain and AI components
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10

# PDF processing
pypdf>=3.17.0

# Vector database
faiss-cpu>=1.7.4

# Machine learning and embeddings
sentence-transformers>=2.2.0

# OpenAI
openai>=1.3.0
tiktoken>=0.5.0

# Scientific computing (pre-compiled wheels)
numpy>=1.26.0
scipy>=1.11.0

# Additional dependencies
requests>=2.31.0
typing-extensions>=4.8.0
python-multipart==0.0.20
